Implementado mapeadores

+++ movies4/src/main/java/com/fpmislata/movies/domain/service/impl/DirectorServiceImpl.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 
 import com.fpmislata.movies.domain.entity.Director;
 import com.fpmislata.movies.domain.service.DirectorService;
-import com.fpmislata.movies.persistence.DirectorRepository;
+import com.fpmislata.movies.domain.service.DirectorRepository;
 
 @Service
 public class DirectorServiceImpl implements DirectorService {
@@ -21,13 +21,24 @@ public class DirectorServiceImpl implements DirectorService {
 
     @Override
     public void update(Director director) {
-        Director existingDirector = directorRepository.find(director.getId()).orElseThrow(() -> new ResourceNotFoundException("Director no encontrado con id: " + director.getId()));
+        Director existingDirector = directorRepository.find(director.getId());
+        if (existingDirector==null){
+            throw new ResourceNotFoundException("Director no encontrado con id: " + director.getId());
+        }
         directorRepository.update(director);
     }
 
     @Override
     public void delete(int id) {
-        Director director = directorRepository.find(id).orElseThrow(() -> new ResourceNotFoundException("Director no encontrado con id: " + id));
+        Director director = directorRepository.find(id);
+        if (director==null){
+            throw new ResourceNotFoundException("Director no encontrado con id: " + id);
+        }
         directorRepository.delete(id);
+    }
+
+    @Override
+    public Director find(int id) {
+        return directorRepository.find(id);
     }
 }
